refactor: renamed entity package to model
