+++ src/main/java/com/henry/universitycourseschedular/controllers/AuthenticationController.java
@@ -2,12 +2,11 @@ package com.henry.universitycourseschedular.controllers;
 
 import com.henry.universitycourseschedular.constants.StatusCodes;
 import com.henry.universitycourseschedular.dto.*;
-import com.henry.universitycourseschedular.entity.AuthToken;
+import com.henry.universitycourseschedular.models.AuthToken;
 import com.henry.universitycourseschedular.repositories.AuthTokenRepository;
 import com.henry.universitycourseschedular.services.AuthenticationService;
 import com.henry.universitycourseschedular.services.OtpService;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
