+++ src/main/java/com/henry/universitycourseschedular/controllers/InvitationController.java
@@ -3,7 +3,7 @@ package com.henry.universitycourseschedular.controllers;
 import com.henry.universitycourseschedular.dto.DefaultApiResponse;
 import com.henry.universitycourseschedular.dto.InviteHodDto;
 import com.henry.universitycourseschedular.dto.SuccessfulInviteDto;
-import com.henry.universitycourseschedular.entity.Invitation;
+import com.henry.universitycourseschedular.models.Invitation;
 import com.henry.universitycourseschedular.services.InvitationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
