+++ src/main/java/com/henry/universitycourseschedular/models/TimeSlot.java
@@ -1,8 +1,8 @@
-package com.henry.universitycourseschedular.entity;
+package com.henry.universitycourseschedular.models;
 
 import jakarta.persistence.*;
 import lombok.*;
-import java.time.LocalDateTime;
+
 import java.time.LocalTime;
 
 @Entity
