+++ src/main/java/com/henry/universitycourseschedular/repositories/AppUserRepository.java
@@ -1,6 +1,6 @@
 package com.henry.universitycourseschedular.repositories;
 
-import com.henry.universitycourseschedular.entity.AppUser;
+import com.henry.universitycourseschedular.models.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
