+++ src/main/java/com/henry/universitycourseschedular/repositories/InvitationRepository.java
@@ -1,6 +1,6 @@
 package com.henry.universitycourseschedular.repositories;
 
-import com.henry.universitycourseschedular.entity.Invitation;
+import com.henry.universitycourseschedular.models.Invitation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
