+++ src/main/java/com/henry/universitycourseschedular/repositories/OtpRepository.java
@@ -1,6 +1,6 @@
 package com.henry.universitycourseschedular.repositories;
 
-import com.henry.universitycourseschedular.entity.OneTimePassword;
+import com.henry.universitycourseschedular.models.OneTimePassword;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
