+++ src/main/java/com/henry/universitycourseschedular/services/OtpService.java
@@ -4,8 +4,8 @@ import com.henry.universitycourseschedular.constants.StatusCodes;
 import com.henry.universitycourseschedular.dto.AppUserDto;
 import com.henry.universitycourseschedular.dto.DefaultApiResponse;
 import com.henry.universitycourseschedular.dto.OneTimePasswordDto;
-import com.henry.universitycourseschedular.entity.AppUser;
-import com.henry.universitycourseschedular.entity.OneTimePassword;
+import com.henry.universitycourseschedular.models.AppUser;
+import com.henry.universitycourseschedular.models.OneTimePassword;
 import com.henry.universitycourseschedular.enums.ContextType;
 import com.henry.universitycourseschedular.enums.VerifyOtpResponse;
 import com.henry.universitycourseschedular.repositories.AppUserRepository;
